+++ apps/RESTBridge.cpp
@@ -3,7 +3,7 @@
  */
 
 #include <restbridge/RestBridge.h>
-#include <lunchbox/log.h>
+#include <restbridge/log.h>
 #include <boost/program_options.hpp>
 
 namespace po = boost::program_options;
@@ -14,16 +14,13 @@ namespace po = boost::program_options;
 
 int main( int argc, char * argv[] )
 {
-    lunchbox::Log::instance().level = lunchbox::LOG_ALL;
-
     // Declare the supported options.
     po::options_description description("Usage:");
     description.add_options()
         ( PARAM_HOST ",h", po::value<std::string>(), "Hostname or IP address")
         ( PARAM_PORT ",p", po::value<std::string>(), "Listening port")
         ( PARAM_SCHEMA ",s", po::value<std::string>(),
-            "Schema prefix used by zeq publisher and subscriber")
-    ;
+            "Schema prefix used by zeq publisher and subscriber");
 
     po::variables_map vm;
     po::store(po::command_line_parser(argc, argv).options(description).run(), vm);
@@ -31,7 +28,7 @@ int main( int argc, char * argv[] )
 
     if( !vm.count( PARAM_HOST ) || !vm.count( PARAM_PORT ) || !vm.count( PARAM_SCHEMA ))
     {
-        LBERROR << description << std::endl;
+        RBERROR << description << std::endl;
         return 1;
     }
 
@@ -45,13 +42,13 @@ int main( int argc, char * argv[] )
     }
     catch( const boost::bad_lexical_cast &e )
     {
-        LBERROR << vm[ PARAM_PORT ].as< std::string >() << " is an invalid port number: " <<
+        RBERROR << vm[ PARAM_PORT ].as< std::string >() << " is an invalid port number: " <<
             e.what() << std::endl;
         return 1;
     }
     catch ( std::exception &e )
     {
-        LBERROR << e.what() << std::endl;
+        RBERROR << e.what() << std::endl;
         return 1;
     }
     return 0;